Rename model files with capital letter
