+++ frontend/src/models/Battery.ts
